Implementadas clases vacías de los paquetes api, controllers y dao.

+++ src/main/java/daos/alarm/AlarmDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import entities.alarm.Alarm;
 
-public interface AlarmDao extends JpaRepository<Alarm, Long> {
+public interface AlarmDao extends JpaRepository<Alarm, Integer> {
 
 }
